+++ data-plane/bindings/rust/src/client_config.rs
@@ -346,6 +346,7 @@ impl From<ClientConfig> for CoreClientConfig {
             require_header_mac: config
                 .require_header_mac
                 .unwrap_or(core_defaults.require_header_mac),
+            connection_type: core_defaults.connection_type,
         }
     }
 }